refactor: Update get function in workspaces.ts to fetch workspaces based on user membership
This commit refactors the get function in workspaces.ts to fetch workspaces based on the user's membership. Previously, it was fetching all workspaces, but now it only returns the workspaces that the user is a member of. The function now retrieves the user's ID using the auth.getUserId() function and then queries the "members" table to get the workspace IDs associated with the user. It then retrieves the corresponding workspaces using the workspace IDs and returns them.